refactor(api/test): migrate to 'socket:test'
