+++ api/test.js
@@ -1,0 +1,3 @@
+import test from './test/index.js'
+export * from './test/index.js'
+export default test
