--- api/test/tap.js
@@ -1,1 +1,0 @@
-export * from './tap/index.js'
