+++ npm/packages/@socketsupply/socket-linux-x64/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@socketsupply/socket-linux-x64",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "A Cross-Platform, Native Runtime for Desktop and Mobile Apps — Create apps using HTML, CSS, and JavaScript. Written from the ground up to be small and maintainable.",
   "type": "module",
   "main": "src/index.js",
