+++ test/src/fs/promises.js
@@ -4,7 +4,7 @@ import fs from 'socket:fs/promises'
 import os from 'socket:os'
 
 import { FileHandle } from 'socket:fs/handle'
-import { test } from 'socket:test/tap'
+import { test } from 'socket:test'
 import { Dir } from 'socket:fs/dir'
 
 const TMPDIR = `${os.tmpdir()}${path.sep}`
