+++ test/src/util.js
@@ -1,5 +1,5 @@
 import util, { hasOwnProperty } from 'socket:util'
-import { test } from 'socket:test/tap'
+import { test } from 'socket:test'
 import Buffer from 'socket:buffer'
 
 test('util.hasOwnProperty', (t) => {
